Revert "moved App.js out of components folder"
