+++ client/package.json
@@ -9,7 +9,6 @@
     "react-card-flip": "^1.0.11",
     "react-dom": "^16.13.1",
     "react-scripts": "^4.0.1",
-    "react-shuffle": "^2.1.0",
     "react-timer-hook": "^2.0.7"
   },
   "scripts": {
