+++ client/src/components/Score.jsx
@@ -8,14 +8,14 @@ function Score() {
           <div className="row">
             <div className="input-field col s6">
               <input id="input_text" type="text" data-length="3"></input>
-              <label htmlFor="input_text">Input Initials</label>
+              <label for="input_text">Input Initials</label>
             </div>
           </div>
 
           <div className="row">
             <div className="input-field col s12">
               <textarea id="input_text2" data-length="12"></textarea>
-              <label htmlFor="input_text2">Score</label>
+              <label for="input_text2">Score</label>
             </div>
           </div>
         </form>
