+++ package.json
@@ -15,7 +15,7 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "concurrently": "^4.1.2",
+    "concurrently": "^4.1.0",
     "nodemon": "^1.18.7"
   },
   "dependencies": {
